fix(Navigation): more fixes on how we handle toggle change
