+++ .changeset/twenty-shrimps-talk.md
@@ -1,0 +1,7 @@
+---
+"@ultraviolet/plus": patch
+---
+
+Some fixes on `<Navigation />`:
+- Rename of `onClickExpand` to `onExpandChange` on `<NavigationProvider />`
+- Add new prop `onToggleExpand` on `<Navigation />` => this function will be triggered on click on the expand/collapse button
