+++ packages/plus/src/components/Navigation/constants.ts
@@ -33,23 +33,3 @@ export const groupAnimation = keyframes`
     margin-bottom: 0;
   }
 `
-
-export type DragNDropData = {
-  index: number
-  item: string
-}
-
-export type PinUnPinType = {
-  /**
-   * The state of the item after the click
-   */
-  state?: 'pin' | 'unpin'
-  /**
-   * The current item id that has been pinned on click
-   */
-  id?: string
-  /**
-   * The total pinned items including the current one
-   */
-  totalPinned?: string[]
-}
